Moved calendar page into separate file.
